moving this stuff to tests
